requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf 
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
lxml
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain_groq
fastapi
uvicorn
langserve